test: fix compilation errors for simplified schema
- Fixed phantom database columns in SQL queries (estimated_completion_at, max_retries, last_retry_at)
- Updated HeartRateMetric test fields from min_bpm/avg_bpm/max_bpm to heart_rate/resting_heart_rate
- Fixed context field type from String to ActivityContext enum in test fixtures
- Added missing UUID, timestamp, and heart_rate_variability fields to test models
- Updated WorkoutType field to use enum instead of String
- Partially resolved test compilation issues in 5 critical test files
- 13 test files still need complete field name updates

Progress: Reduced critical test compilation failures, 156 errors remain

+++ src/models/health_metrics.rs
@@ -341,7 +341,7 @@ impl SleepMetric {
         }
 
         if let Some(eff) = self.efficiency {
-            if !(config.sleep_efficiency_min..=config.sleep_efficiency_max).contains(&eff) {
+            if !(config.sleep_efficiency_min..=config.sleep_efficiency_max).contains(&(eff as f32)) {
                 return Err(format!(
                     "efficiency {} is out of range ({}-{})",
                     eff, config.sleep_efficiency_min, config.sleep_efficiency_max
